requirements.txt
Flask
line-bot-sdk
uwsgi